+++ index.html
@@ -10,7 +10,10 @@
 </head>
 
 <body>
-  <script src="js/main.js"></script>
+    <script src="js/main.js" type="module"></script>
+    <script src="js/module/util.js" type="module"></script>
+    <script src="js/module/data.js" type="module"></script>
+
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
